+++ test_remote.ts
@@ -3,33 +3,6 @@ import { prepare, PerpareOptions } from "./mod.ts";
 
 const textDecoder = new TextDecoder();
 
-async function testPrepare() {
-  const releaseUrl =
-    "https://github.com/eliassjogreen/deno-plugin-prepare/releases/download/plugin_bins";
-
-  const pluginOptions: PerpareOptions = {
-    name: "test_plugin",
-    printLog: true,
-    urls: {
-      darwin: `${releaseUrl}/libtest_plugin.dylib`,
-      windows: `${releaseUrl}/test_plugin.dll`,
-      linux: `${releaseUrl}/libtest_plugin.so`,
-    },
-  };
-  const pluginId = await prepare(pluginOptions);
-  //@ts-ignore
-  const { testSync } = Deno.core.ops();
-
-  //@ts-ignore
-  const response = Deno.core.dispatch(
-    testSync,
-    new Uint8Array([116, 101, 115, 116]),
-    new Uint8Array([116, 101, 115, 116]),
-  )!;
-
-  assertEquals(textDecoder.decode(response), "test");
-}
-
 async function testPrepareFromLocal() {
   const releaseUrl = "file://./test_bins";
 
@@ -53,8 +26,8 @@ async function testPrepareFromLocal() {
     new Uint8Array([116, 101, 115, 116]),
     new Uint8Array([116, 101, 115, 116]),
   )!;
+  
   assertEquals(textDecoder.decode(response), "test");
 }
 
-await testPrepare();
 await testPrepareFromLocal();
